stun: fix build issue

+++ link/stun/src/stun/rfc5766.go
@@ -895,7 +895,7 @@ func getChannelData(buf []byte) (*channelData, error) {
 
 	// there may be paddings according to https://tools.ietf.org/html/rfc5766#section-11.5
 	length := binary.BigEndian.Uint16(buf[2:])
-	if 4 + length > len(buf) {
+	if 4 + length > uint16(len(buf)) {
 		return nil, fmt.Errorf("channelData length is too large")
 	}
